requirements.txt
# LangGraph and LangChain dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-core>=0.3.0

# For type hints
typing-extensions>=4.0.0

# Environment variables
python-dotenv>=1.0.0
